Adjust build files for VPATH builds (recommended)
In order to keep the source directories clean, VPATH builds are from
now on favored. This means that the build tree is kept separate from
the source tree, such that the built files are not in the same place
as the source files. For this to work, several paths in the
`Makefile.am' files had to be adjusted.