Introduce new tilt axis determination
The RecParams structure contains a new 'tilt_axis' member.
If the tilt_axis parameter in the config file is between -45 and +45 degrees,
tilt_axis is set to 0 (-> x axis tilt) and 'tilt_axis_rotation' is set to the
config file parameter.
Otherwise, tilt_axis is set to 1 (-> y axis tilt), and 'tilt_axis_rotation'
is calculated as the config file parameter minus (-)90 degrees according to
the sign of the parameter.
This way, image_rotation will automatically align the tilt axis with either
the x or y axis, and the correct backprojection function will be called.